electron loading update
updated the loading process for electron to use `main.html` vs relying
on a generated file.

the deprecated packages that where showing during the install process
seemed to be coming from `react-scripts` dependencies. we now will now
use a more manual process to build the files and will be using `rollup`
to bundle everything needed.

I also updated the `README` with some information about how to build and
run the electron instance.